[FE] some admin dashboard fix
